fix(security): decimal refactor — float → Decimal for monetary values
PR #5 of 6-PR security remediation plan (audit 2026-06-15). F-002 (CRITICAL).

- utils/money.py: Money type alias (Decimal 12/2), to_decimal128, from_decimal128, quantize_money (ROUND_HALF_UP), decimalize_doc
- models.py: 24 monetary float fields → Money (Decimal)
- 8 services: Decimal arithmetic with ROUND_HALF_UP
- routers/admin.py: revenue $sum, bulk-update, shipping Query → Decimal
- routers/products.py: min_price/max_price → Decimal
- email_service.py: Decimal signature
- scripts/migrate_floats_to_decimal128.py: idempotent migration + --downgrade
- 259 tests passing, 3 xfailed (mongomock Decimal128 sort limitation)

2 tasks pending external artifacts: MANUAL_SMOKE.md, staging validation.

+++ routers/products.py
@@ -1,5 +1,6 @@
 from fastapi import APIRouter, Depends, HTTPException, status, Query, Response
 from typing import Optional
+from decimal import Decimal
 from bson import ObjectId
 
 from motor.motor_asyncio import AsyncIOMotorDatabase
@@ -39,8 +40,8 @@ async def create_product(
 async def read_products(
     db: AsyncIOMotorDatabase = Depends(get_database),
     category: Optional[ProductCategory] = Query(None, description="Filtrar por categoría de producto"),
-    min_price: Optional[float] = Query(None, ge=0, description="Precio mínimo del producto"),
-    max_price: Optional[float] = Query(None, ge=0, description="Precio máximo del producto"),
+    min_price: Optional[Decimal] = Query(None, ge=0, description="Precio mínimo del producto"),
+    max_price: Optional[Decimal] = Query(None, ge=0, description="Precio máximo del producto"),
     search: Optional[str] = Query(None, min_length=2, description="Buscar por nombre o descripción del producto"),
     include_out_of_stock: bool = Query(False, description="Incluir productos sin stock (para administradores)"),
     page: int = Query(1, ge=1, description="Número de página (1-indexed)"),
